Stop using env in python shebang
This change is motivated by
https://stackoverflow.com/a/25338292/615168,
https://stackoverflow.com/q/6495688/615168 and
https://unix.stackexchange.com/q/29608/99614 with the answer by Keith
Thompson.

Basicly there is a tradeoff, env does handle some edge cases but
introduces some new ones. Nowadays the assumption that python3 or a
symlink to it will be installed under /usr/bin/python3 is rather
safe. If it still turns out to be false for someone, one can always
add a symlink or invoke python3 directly.